+++ metadata/changelog.md
@@ -15,6 +15,7 @@
 - Buffed Thermal Expansion's Dynamo Generation.
 - Blacklisted Evilcraft's Display Stands and Biome Bottles from HEI.
 - Added a Mixin to Fix an Issue with Embers Freeze during Boot.
+- Reworked Immersive Engineering's Blueprint Recipes.
 
 ---